Split config models and extract composition helpers
