+++ src/yoyopod/config/models/__init__.py
@@ -1,0 +1,87 @@
+"""Typed configuration models and YAML loading helpers."""
+
+from yoyopod.config.models.app import (
+    AppDiagnosticsConfig,
+    AppDisplayConfig,
+    AppInputConfig,
+    AppLoggingConfig,
+    AppMetadataConfig,
+    AppUiConfig,
+    YoyoPodConfig,
+)
+from yoyopod.config.models.cloud import (
+    BackendTelemetryConfig,
+    CloudBackendConfig,
+    CloudConfig,
+    CloudSecretsConfig,
+)
+from yoyopod.config.models.communication import (
+    CommunicationAccountConfig,
+    CommunicationAudioConfig,
+    CommunicationCallingConfig,
+    CommunicationConfig,
+    CommunicationIntegrationsConfig,
+    CommunicationMessagingConfig,
+    CommunicationNetworkConfig,
+    CommunicationSecretConfig,
+)
+from yoyopod.config.models.core import (
+    build_config_model,
+    config_to_dict,
+    config_value,
+    load_config_model_from_yaml,
+)
+from yoyopod.config.models.media import MediaAudioConfig, MediaConfig, MediaMusicConfig
+from yoyopod.config.models.network import NetworkConfig
+from yoyopod.config.models.people import PeopleDirectoryConfig
+from yoyopod.config.models.power import (
+    GpioPin,
+    PimoroniGpioConfig,
+    PimoroniGpioInputConfig,
+    PowerConfig,
+)
+from yoyopod.config.models.runtime import YoyoPodRuntimeConfig
+from yoyopod.config.models.voice import (
+    VoiceAssistantConfig,
+    VoiceAudioConfig,
+    VoiceConfig,
+)
+
+__all__ = [
+    "AppDiagnosticsConfig",
+    "AppDisplayConfig",
+    "AppInputConfig",
+    "AppLoggingConfig",
+    "AppMetadataConfig",
+    "AppUiConfig",
+    "BackendTelemetryConfig",
+    "CloudBackendConfig",
+    "CloudConfig",
+    "CloudSecretsConfig",
+    "CommunicationAccountConfig",
+    "CommunicationAudioConfig",
+    "CommunicationCallingConfig",
+    "CommunicationConfig",
+    "CommunicationIntegrationsConfig",
+    "CommunicationMessagingConfig",
+    "CommunicationNetworkConfig",
+    "CommunicationSecretConfig",
+    "GpioPin",
+    "MediaAudioConfig",
+    "MediaConfig",
+    "MediaMusicConfig",
+    "NetworkConfig",
+    "PeopleDirectoryConfig",
+    "PimoroniGpioConfig",
+    "PimoroniGpioInputConfig",
+    "PowerConfig",
+    "VoiceAssistantConfig",
+    "VoiceAudioConfig",
+    "VoiceConfig",
+    "YoyoPodConfig",
+    "YoyoPodRuntimeConfig",
+    "build_config_model",
+    "config_to_dict",
+    "config_value",
+    "load_config_model_from_yaml",
+]
